Only trigger a datafusion update when the relevant part of the blueprint changes (#9916)
### Related

* in prep for https://github.com/rerun-io/rerun/issues/9795

### What

Another pure refactor to express the fact that only _part_ of a
`TableBlueprint` influences/is handled by the datafusion query.
Currently, that "part" is 100% of the table blueprint, but I want to add
new stuff to it that does not concern datafusion. This PR makes this
separation of concern more explicit.