redap: support REDAP_TOKEN env var (#9935)
### What
To access redap endpoints that require authentication (such as
`sandbox.redap.rerun.io`) we need ways to pass the auth token to the
GRPC client.

This is the simplest most naive way, that unblocks us.

We may consider to have explicit SDK params and even UI elements in the
viewer for that.

### Notable miss
This won't work with with the `wasm` build because `re_auth` is not
friendly with it. I have not had the time to dig into it, and I will
likely need help with that. Created an issue in
https://github.com/rerun-io/dataplatform/issues/726

### Issues
* Closes: https://github.com/rerun-io/dataplatform/issues/722
* Opens :sweat_smile: :
https://github.com/rerun-io/dataplatform/issues/726

---------

Signed-off-by: Andrea Reale <andrea@rerun.io>