Use weaker version of `ComponentDescriptor` for indicators (#9990)
### Related

* Part of #6889.
* From
https://github.com/rerun-io/rerun/pull/9938#issuecomment-2888808593:

> was curious and poked a lil bit into the ci failure: took me a while
to notice but the primary issue presented by the failing images is that
it draws lines instead of points.
> 
> Which implies an issue with indicators (wtf why is that the only test
failing then.. but sure usually there's no choice; have only a vague
hunch about that :/). Looking at `impl<A: Archetype>
crate::ComponentBatch for GenericIndicatorComponentArray<A>` indicators
are never logged with their full descriptor, but are advertised as such.
> 
> So I quickly hacked the generated `SeriesPoints::descriptor_indicator`
to be name only and sure enough that fixed the test. What to do about
this knowledge? Uhm.. yeah... that is definitely a Monday problem 😄

### What

Because indicators are about to go away soon (™️) anyways, making them
weaker (qualification-wise) is probably the easiest choice.

Doing this in a separate PR to help with potential `git bisect` in the
future.