tasks: wait_for_tasks can wait for streaming responses (#10015)
### Related
* Required by: https://github.com/rerun-io/dataplatform/pull/762

### What

In https://github.com/rerun-io/dataplatform/pull/762 we made the
tasks-completion streaming endpoitn a real streaming endpoint; before it
was a stream with one item.

This PR adusts the `wait_for_tasks` endpoint so that it does no longer
make the assumption there is only one item ever in the stream.

### Testing

I've registered a droid50 on my local redap server using
`register_batch` from the SDK. I tested with both the previous and the
current server version (the thing should be backwards compatible).

### What next
We could go further and improve the API adding a different
`subscribe-tasks` method that yields results as tasks complete. This
would let higher level utils like the dataset-util do a better job at a
progress bar. However, I did not adventure that far, as this was for me
just a way to unblock 762.

---------

Signed-off-by: Andrea Reale <andrea@rerun.io>