Introduce variant UIs in `re_component_ui` (#10034)
### Related

* Part of https://github.com/rerun-io/rerun/issues/9741
* Part of https://github.com/rerun-io/rerun/issues/9795

### What

This PR add the capability to register and use arbitrary pieces of
Arrow-based UI that are not necessarily linked to a specific component.
They are called "variant UI", as opposed to "component UI".

To summarise, **component UIs**:
- are registered against a component name
- are (typically) implemented with closures that accept (and return, if
editable) fully deserialised component
- (because of what precedes) are inflexible w.r.t the arrow datatype
then can handle
- support view and edit, single line and multiline

In contrast, **variant UIs** (for now):
- are registered against a `VariantName` (conceptually a `&'static str`
constant)
- are implemented with closures that accept raw arrow data
- are flexible w.r.t the arrow datatype they can handle
- (because of what precedes) should explicitly document what arrow
datatype they expect/support
- only support edit, single line (for now)

Implementation notes:
- For now, the variant UI are registered with `add_variant_ui` and used
with `variant_ui_raw`, which is mostly an independent, parallel path
from the pre-existing stuff.
- Component descriptor and row id are now passed to
`UntypedComponentEditOrViewCallback`, because these ~are~ will be
convenient to have in variant ui impls.
- This PR includes a `redap_uri_button` variant ui, which handles data
that happens to be rerun-conforming URIs.