Add support for variant UIs in the DataFusion table widget and show a button in the partition table recording links (#10035)
### Related

- fixes #9795 
- follow-up to https://github.com/rerun-io/rerun/pull/10034
- part of #9741 
- future work https://github.com/rerun-io/rerun/issues/10036

### What

This PR introduces the ability to specify on a per-column basis a
specific variant ui (see #10034) to use. To that end, a new
`ColumnBlueprint` structure is introduced and a bunch of things around
that is refactored.

This PR also uses the recently introduced `redap_url` variant UI to
display the recording link as an "open" button instead of a fully URL.



![image](https://github.com/user-attachments/assets/d512d62c-ac3c-4c46-9caa-4c9bc5305d4f)