Improve UX of opening recordings in catalog browser (#10080)
### Related
* Closes https://github.com/rerun-io/rerun/issues/10067
* Closes https://github.com/rerun-io/rerun/issues/10066

### What
#### Fix "open in background" when clicking redap recording links
This adds a `bool` flag to the `re_smart_channel` data srouce to
indicate wether or not the user intended to switch to this recording
once it has loaded.

Putting it in the smart channel was not the easiest decision, but I
think it was the least bad.

For one, the smart channel source already keeps track of things like the
URI fragment which already contains instructions on what to do next,
once the URI is loaded (e.g. go to a specific timeline).

But also, I found no better place for this data.

#### Show "Switch to" button if recording is already open
cmd-clicking a few links:

![open-in-background](https://github.com/user-attachments/assets/a01143d2-0901-4ef7-828f-37e9d0ed60b5)