Do not hold on to the `RedapClient` in the SDK (#10097)
### Related

- Follow up to #10078 

### What

One should not hold to a `RedapClient` for extended periods of time, as
they might need to be periodically refreshed in the future (for auth
purposes). This PR changes the SDK implementation such that it always
gets a fresh copy of the client instead of holding on to the same one
forever.