separate wasm binary from js widget
The `rerun_notebook` build has been updated to:
* Copy the Wasm binary from node_modules to `/static`
* Preserve special inline marker comments

The asset server has been updated to also serve the `re_viewer_bg.wasm`
file. It serves it gzipped, with the right `content-type` so that it can
be stream-compiled.

The `rerun-js/web-viewer` build no longer combines the Wasm and JS into
a single file - this is instead done at _runtime_, by the
`rerun_notebook` package, becuase we don't want to publish two copies of
the Wasm binary (one as its own file, one inlined into the JS widget)
when publishing the package. It has a negligible startup cost, which is
paid once at time of `import rerun.notebook`, and only if
`RERUN_NOTEBOOK_ASSET` is set to `inline`.