`block_until_ready=False` by default
The problem this was attempting to solve was the widget asset not being
accessible for whatever reason causing the widget to never finish
loading without any error messages, unless you opened the browser
devtools console. That was a super frustrating experience, so we instead
set a hard limit (10 seconds) for how long the Viewer should take to
load. That turns out to just not be enough in many environments...

Because we now check if the required assets are accessible during import
time, we no longer need this, as we won't ever timeout in the same way.