Use egui atoms in help view (#10224)
### Related

- atomics landed in egui (https://github.com/emilk/egui/pull/5830)

### What

Updates egui to latest master and uses atoms in the help view, removing
the icon_text macro since it was basically atomics lite

Seems like the snapshots the only thing that changed is the three dot
icon slightly moved. But it seems a bit clearer now, so that's a win.