Add duration field to the recording UI (#10284)
This is just a small QoL improvement. I found myself looking for this
many times when comparing recordings. In particular, I wanted to see the
duration of a recording to help compare the sizes between different
recordings (I'm evaluating the storage saved by using video).

If the recording has a `log_time` timeline, it will display a human-readable
format like "2m 12.34s" to 2 levels of precision on the seconds field.