Create a globally accessible analytics object (#10333)
### Related
* https://github.com/rerun-io/rerun/pull/10301

### What
In https://github.com/rerun-io/rerun/pull/10301 I want to send an
analytics event the first time a user clicks a help-button… but there is
no way to access the `Analytics` object deep down in `re_ui` without a
lot of ugly plumbing. So this PR makes `Analytics` globally accessible
so that we can send analytics from anywhere.

In the crash handler we now also use the global analytics object,
instead of creating a new one (unless a new one is needed). This makes
the analytics handling in the crash handler slightly cleaner and faster.