Improve migration code (#10365)
This cleans up the sorbet migration code in several ways (see commit
messages).

It also fixes a footgun: we would sometimes try to parse just the batch
schema, without for migrating it. This is now forbidden. All parsing of
sorbet data must be done on the batch-level, so that it is properly
migrated.

If this becomes a problem, let me know.

### TODO
* [x] `@rerun-bot full-check` (failed with the same error as `main`)
* [x] Check dataplatform
   * [x] still compiles
   * [x] Passes `cargo test`