Fix migration of indicators from column fields (#10375)
### What

Indicators that are present on the `field.name()` were not properly
migrated to `rerun:component` as they were missing the
`rerun.components.` prefix. (💩)

It was not clear from the output of our compare command because all
sorts of prefixes were trimmed. This PR also adds options to not trim
anything to our chunk formatter. So, hopefully, no one will ever have to
go through this again.

* [x] full build

---------

Co-authored-by: Emil Ernerfeldt <emil.ernerfeldt@gmail.com>