Auto render juypyter notebooks with rerun specified formatter (#10330)
### Related

https://linear.app/rerun/issue/DPF-1796/niko-feedback-triage-1

### What

This PR adds in a simpler html formatter for the jupyter notebook
displays. It also leverages the existing formatting code in
`re_format_arrow` so that we truncate data in the same way in jupyter
notebooks as we do in the terminal.

Attached is a screenshot of the updated table display.

<img width="1057" alt="Screenshot 2025-06-23 at 9 37 42 PM"
src="https://github.com/user-attachments/assets/9227e8d1-095c-4dde-9b70-49fe84138e23"
/>

---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>