Replace `rerun:version` with `sorbet:version` to help with future migrations (#10392)
### Related

* Part of #8748.
* Closes #10322.

### What

This PR replaces the `rerun:version` chunk metadata with a
`sorbet:version` metadata that can be used to inform migrations, but
crucially is decoupled from the Rerun version to avoid confusion.

The Sorbet version should follow semantic versioning and will be used to
determine which migrations to apply.

We start with `sorbet:version: 0.1.0` until we hit a stable state.
Previous migrations are mapped to this scheme (`0.0.1` and `0.0.2`) for
consistency.

### Chunk store

<img width="476" alt="image"
src="https://github.com/user-attachments/assets/490db085-8c04-4ed9-87f7-0381dd51724b"
/>

### TODO

* [x] full check
* [x] Dataplatform companion PR:
https://github.com/rerun-io/dataplatform/pull/1059