Show a blinking orange dot in debug builds on multi-pass egui frame (#10434)
### Related
* https://github.com/rerun-io/rerun/pull/10406
* https://github.com/emilk/egui/pull/7276

### What

Show an orange dot to warn about multi-pass layout in egui.

If it is shown, it means something called
`egui::Context::request_discard` the previous pass,
causing a multi-pass layout frame in egui.
This is used to cover up some visual glitches, but it is also
a bit costly and we shouldn't do it too often.

An infrequent blinking of the dot (e.g. when opening a new panel) is
expected,
but it should not be sustained.



![blink-dot](https://github.com/user-attachments/assets/599cd46e-e087-44a3-9153-b746005df319)

In this example the selection panel (outside the frame) shows a
`ListItem` scope for the first time, which results in a single
multi-pass frame.