Fix videos re-decoding on some ui interactions (#10420)
### Related

* Fixes https://github.com/rerun-io/rerun/issues/10295

### What
This properly evicts videos only if they weren't in used in _any pass_
during the previous _frame_.

In egui, each frame is _usually_ one pass, but can be two (if something
calls `egui::Context::request_discard`). In that case, the video may be
invisible one pass (because we early-out), but visible the next. Before
that lead to an unecessary eviction. The new code solves this.

---------

Co-authored-by: Emil Ernerfeldt <emil.ernerfeldt@gmail.com>