Move `fuzzy_descriptors` up into the request object, with its peers (#10482)
This just moves `fuzzy_descriptors` where it should have been to start
with, to prevent the introduction of yet more hacks.

We already know that all the protobuf Query objects need to be
redesigned from scratch anyhow, but this took very little effort and
removes yet another layer of insanity.

* Related to https://github.com/rerun-io/dataplatform/issues/1123
* Sibling: https://github.com/rerun-io/dataplatform/pull/1128