Add formatted str representation for PySchema (#10428)
### Related
Related to #9457 but dramatically less ambitious. Right now we have no
repr or str so it doesn't print out anything useful.

### What
This adds a basic __str__ representation. Since this is often getting
used in notebooks I don't know if we have a hardline policy on repr vs
str. Repr should allow us to recreate the object with the information,
but that is what gets displayed if the variable name is typed in a
standalone cell in Jupyter. Str is allowed to be more pretty print.
Since we didn't have anything before I did a pretty print and set the
repr to that.

Example output
```python
>>> dataset.schema()
Column name: /action/gripper_position:Scalars:scalars
                  Entity path: /action/gripper_position
                  Archetype type: rerun.archetypes.Scalars
                  Component type: rerun.components.Scalar
<repeat for ALL columns>
>>> for field in dataset.schema():
...if "link" not in field.entity_path and not in field.component.is_indicator():
......print(field)
Column name: /action/gripper_position:Scalars:scalars
                  Entity path: /action/gripper_position
                  Archetype type: rerun.archetypes.Scalars
                  Component type: rerun.components.Scalar
<repeat for columns not link or indicator>
```

---------

Co-authored-by: Gijs de Jong <14833076+oxkitsune@users.noreply.github.com>