Stabilize light mode (#10500)
### Related
* Closes https://github.com/rerun-io/rerun/issues/10289
* Closes #3058 

### What
This is one of the last remaining blockers for light mode!

This affects any part of the UI where we use blue to indicate
"selection", including:
* `ListItem`:s (e.g. blueprint panel tree)
* Selection panel breadcrumbs
* Selected tabs in viewport
* Play-button

The difference is largest in light mode.


![image](https://github.com/user-attachments/assets/8a47670e-c8ab-4f3d-a90b-97bbbed2728e)

![image](https://github.com/user-attachments/assets/7460cdfa-bec2-4dd5-ac30-03ba328e74d5)

![image](https://github.com/user-attachments/assets/876faf6d-8d8c-462b-80e5-ce9b8ab2475b)

![image](https://github.com/user-attachments/assets/08873361-d014-49c3-a7b2-dcd17b7e9c86)


### Other

In addition, I also made sure the background of the `>` collapsing icon
gets filled when hovered:


![image](https://github.com/user-attachments/assets/1664ef0e-ec90-4bb2-bc92-8be7bd9ff676)


### Other other
This also enables switching to light mode in release builds! This means
light mode is now OFFICIALLY SUPPORTED