Fix: don't always pick `/` as the origin when creating new views (#10514)
### What
When creating a new view from a selection of entities (e.g. by
right-clicking them and selecting "Add to new view") the origin would
always be set to `/` (except for 2D and 3D views). This was quite
confusing, at it also meant the new view would be given the name `/`.

Review commit-by-commit!
The first commit is the fix, the second just a naming change.


![image](https://github.com/user-attachments/assets/c367df2e-9eeb-44b5-9301-5e42fa28ae78)