Free up memory of inactive recordings (#10537)
### Related
* Related to https://github.com/rerun-io/rerun/issues/10496
* Problems added in https://github.com/rerun-io/rerun/pull/7513

### What
When we want to free up memory, we forgot about caches for other
recordings than the current one.
Similarly, each frame we maintain the caches by throwing out unnecessary
parts of it, but again we only did it for the currently open recording.
This meant recordings that are "in the background" would still use up a
lot of memory.