Removes remnants of indicators from dataframe queries and viewer (#10584)
### Related

* Part of #6889.
* Closes #8129.
* [x] Merge after #10583
* [x] Merge after #10581

### What

This removes the remaining reference to indicators from our codebase.
Most notably this removes indicators from the dataframe queries and the
UI.

This should be fine because:
* We drop indicator columns during Sorbet migration
* By default indicators where not shown anyways
* And they were omitted in queries by default too.

> [!IMPORTANT]
> There is some risk that this PR will make indicators appear again, so
we should choose a good time to merge this PR. If they show up again
then let's fix the root cause (i.e. migration logic).

### TODO

* [x] full-check
* [x] Dataplatform companion PR:
https://github.com/rerun-io/dataplatform/pull/1234