`rerun rrd route` command to manipulate `.rrd`s on transport level (#10607)
### Related

* CLI part of #8860.
* Supersedes #10596.

### What

This adds a new `rerun rrd route` command that can be used to route one
or multiple `.rrd`s to a new one (via `--output`) or to stdout.

This provides a concise way to merge recordings using the CLI. In future
versions, we also want to add that functionality to the UI.

> [!IMPORTANT]
> The `route` command will never decode the Arrow data, therefore no
migration or verification will happen, which is good for performance. If
you still want migration, merging, and compaction, the result can be
piped into `rerun rrd merge`.

Here is an example of the interface, which pipes directly to the viewer:

```sh
rerun rrd route \
  --application-id <new-app-id> \
  --recording-id <new-recording-id> \
  a.rrd b.rrd | rerun -
```

Or you can write the `.rrd` back out:

```sh
rerun rrd route \
  --application-id <new-app-id> \
  --recording-id <new-recording-id> \
  --output <result.rrd> \
  a.rrd b.rrd
```