AnyValue Torch Performance Improvement (#10647)
### Related
Closes #10556 and maybe #10344

### What
* Adds support for the [`__dlpack__`
](https://dmlc.github.io/dlpack/latest/) interface so we can optimally
convert into 1d arrays (ND-Array support still not available and
probably can't do it no copy #4572). This is mostly to support `torch`
but should work for a variety of things.

* Catches the edge cases:
* Calling AnyValues without a kwarg should say SOMETHING, since the
positional only value doesn't generate content for the AnyValue
* Clarify the error message when we hit the fall through for numpy.
Raises from the exception from arrow so we have the full traceback but
tries to provide more guidance when we expect a value but get another.
   
 ### Performance comparison
This is not a strict benchmark but should be sufficiently convincing
things work nice.
 ```python
>>> pixi run -e py python sample.py raw
Loop 100
1
raw:
        0.0024s
>>> pixi run -e py python sample.py to_numpy
Loop 100
1
to_numpy:
        0.0024s
>>> pixi run -e py python sample.py raw_numpy
Loop 100
1
to_numpy:
        0.0025s
```