Improve error handling in `air_traffic_data.py` (#10715)
### What

We are experiencing regular [CI failures on
`nightly`](https://github.com/rerun-io/rerun/actions/runs/16408059984/job/46358076542)
for the Air Traffic Example, which we suspect to be related to
downloading the required data.

This PR adds some basic error handling to catch common HTTP problems, to
report errors when they happen as opposed to poisoning data downstream.