Remove `testing` feature flag from `re_viewer_context` and `re_viewport` (#10712)
### What

This removes the `testing` feature flag from `re_viewer_context` and
`re_viewport` and replaces it with two crates: `re_test_context` and
`re_test_viewport`.

This reduces the amount of feature combinations that cargo has to chew
through and is a first step of cleaning up our test harness for snapshot
and image comparison tests.