Fix the `Partition(Type)` vs. `DataSource(Kind)` situation (#10768)
We have this old problem that the idea of a data source exists twice:
once in the form of `DataSource` (+ kind) and once in the form of
`PartitionDescriptor` (+ kind). Both types are pretty much the exact
same, except with different quirks :unamused:

This was already hurting code comprehension before, but now it's just
straight up blocking: I'm not gonna implement the notion of layering
twice.

* Sibling: https://github.com/rerun-io/dataplatform/pull/1385