Try to Fix Python Typing (#10704)
### Related
Our typing doesn't do what we want. We are running mypy in our base
environment that doesn't have `rerun` installed so we can't properly
check all of our transitive types. See
https://github.com/rerun-io/rerun/pull/10690 and
https://github.com/rerun-io/rerun/issues/10695 for a few footguns.

### What
Updates `py-lint` to be two targets`py-lint-non-sdk` which is now a
separate `.ini` file since we care about that differently, and
`py-lint-rerun` we still run in default environment so imperfect
coverage but we catch more now.

I also went through and "fixed" the broken pieces. Somewhat haphazard
combination of casts, bug fixes, and typing updates.

NOTES:
1. `pixi run -e py py-lint-rerun` isn't integrated anywhere yet. Unclear
where in our CI stack this goes.
2. This will help here but is not a pre-requisite
https://github.com/rerun-io/rerun/issues/10695 (however the combination
of both may complicate things if we go the route of optional
dependencies and testing)
3. General question around BoolArrayLike, FloatArrayLike, is there a
reason for a single value not wrapped in array we don't take the python
equivalent too?