Add some analytics events (#10793)
* Closes https://github.com/rerun-io/rerun/issues/10597

button.

Why? Because we've discovered in user interviews that a lot of people
don't know how to use the viewer, because they never clicked the `?`
question marks that we've scattered around. We've made them a bit more
obvious now, but we still don't know if everyone finds them.

How many users know we have a settings screen?

Collect the native pixel ratio and zoom level to figure out if it is
worth designing Rerun for low-DPI screens