Add some analytics events (#10793)
### Related
* Closes https://github.com/rerun-io/rerun/issues/10597

### What

### Send an analytics event the first time the user clicks an help
button.

Why? Because we've discovered in user interviews that a lot of people
don't know how to use the viewer, because they never clicked the `?`
question marks that we've scattered around. We've made them a bit more
obvious now, but we still don't know if everyone finds them.

### Send analytics when the user enters settings
How many users know we have a settings screen?

### Send analytics about screen scaling
Collect the native pixel ratio and zoom level to figure out if it is
worth designing Rerun for low-DPI screens