Add support for layers to the python registration APIs (#10795)
The gRPC API was already exposed, so this is fairly light.

I went with a clamping-semantics-like approach for the batched
registration API, i.e. we will either truncate or extend the list of
layers automatically as needed. As I implemented a layer-aware
`dataset-util`, this really felt like the most pleasant API to work
with.

* Said `dataset-util` implementation:
https://github.com/rerun-io/dataplatform/pull/1411