Crate publishing: after publishing a crate, wait for it to be available (#10820)
I had this problem:
<img width="865" height="804" alt="image"
src="https://github.com/user-attachments/assets/f4e635d8-eaf6-4fdc-b21f-ea90747a60f0"
/>

So `re_build_info` was successfully published, but when trying to
publish something that was using it, we still failed.
I had a similar problem earlier, but with `re_log`

So I added a loop to wait for the crate to become "visible". I hope that
will help.