Add automatic undo-bug-detection code (#10853)
### Related
* https://github.com/rerun-io/rerun/issues/10304
* https://github.com/rerun-io/rerun/pull/10355

### What
We previously had a bug where we would write to the blueprint store
every frame, which in turn made undo break, silently.

This PR makes the breakage loud (warn-log SPAM) in debug builds, which
should allow us to catch this bug much earlier next time.

To test this I reverted the fix for the previous bug
(https://github.com/rerun-io/rerun/pull/10355) and indeed, the detector
works.

I've done a bit of testing of our UI, and I've found nothing on `main`
that triggers the new warning.

---------

Co-authored-by: Jochen Görtler <grtlr@users.noreply.github.com>