Implement row filtering on client side (#10829)
### Related

Closes https://github.com/rerun-io/dataplatform/issues/851

### What

This PR identifies when the push down filters provided to the table
provider are checking for a component that is not null. If so it pushes
this filter down into the chunk store query. This reduces the amount of
data coming out of the cpu worker thread and passed into the rest of the
datafusion engine.

These are equivalent

- column is not null
- not (column is null)
- column != lit(null)