Check if our CLI docs match `rerun man` (#10864)
### Related

* #10641
* Closes #10640 

### What

This adds a new script that checks if the CLI docs are up to date with
the output of `rerun man` (which we use to generate the docs).

You can try it out via:

```sh
pixi run python scripts/ci/check-cli-docs.py
```

The script (written by Claude) accepts a command to run as an optional
argument so that we can run it in CI with minimal impact (no
rebuilding).

### TODO:

* [x] Based-on: #10856 (which should introduce an error).