Shrink many of our error types (#10845)
Large errors leads to large `Result`s, which may lead to a lot of
`memcpy`:s when returning a `Result` (even an `Ok` one), which can hurt
performance.

Some of our error types were over 100 bytes, even 200 bytes.

Many of our errors are still pretty huge (64+ bytes), but this is a
start.