Support 2GiB+ videos: Encode `Blob` as arrow `Binary` (#10875)
This changes the encoding of the component column containing bobs from
`List<List<u8>>` to `List<Binary>`, i.e. `BinaryArray` or
`LargeBinaryArray` (both are supported!).

Old data is migrated on load.

This adds support for blobs larger than 2 GiB, e.g. large video files.

In the viewer we now support both 32-bit and 64-bit offsets for the
binary data (`BinaryArray vs LargeBinaryArray`).
This is _technically_ a type of datatype generics. Sort of. See
https://github.com/rerun-io/rerun/issues/9144

* The SDK code (Rust, Python, C++) always encoded as 64-bit (supporting
>2GiB videos).
* Legacy blobs are converted from 32-bit List<u8> to 64-bit binary

## TODO
* [x] Fix python
* [x] Merge https://github.com/rerun-io/rerun/pull/10899
* [x] Merge https://github.com/rerun-io/rerun/pull/10898
* [x] Use 63-bit `LargeBinary` instead of 31-bit `Binary`
  * Closes https://github.com/rerun-io/rerun/issues/10929
  * Closes https://github.com/rerun-io/rerun/issues/10973
* [x] `@rerun-bot full-check`