Add test for empty batch creation related to calling count() (#11006)
### Related

Closes https://github.com/rerun-io/rerun/issues/10894

### What

This adds a unit test to verify that we have coverage for cases where we
return empty record batches and need to initialize them with options to
avoid a panic.

I reverted the correction in
https://github.com/rerun-io/rerun/pull/10895 and ran the pytest. It
generates the following error:

```
    def count(self) -> int:
        """Return the total number of rows in this :py:class:`DataFrame`.

        Note that this method will actually run a plan to calculate the
        count, which may be slow for large or complicated DataFrames.

        Returns:
            Number of rows in the DataFrame.
        """
>       return self.df.count()
               ^^^^^^^^^^^^^^^
E       Exception: DataFusion error: Execution("Arrow error: Invalid argument error: must either specify a row count or at least one column")
```

With the correction in place, this test passes.