use time ceiling instead of cast in UDF (#11022)
When using nanosecond time stamps that are not easily coercible into
microseconds, the `partition_url_with_timeref_udf` function will throw
an exception. Instead we use the ceiling to convert to nearest
microseconds.

The rest of this PR is improving the unit tests. Specifically:

- Add 20 files to the dataset instead of 2. This tests datafusion's
partitioning scheme better because even with the default 14 threads we
are guaranteed to get more than one partition in a single thread.
- Guarantees we are using timestamps that are not coercible to
microseconds.