Upgrade Ruff to 0.12.10 (#10996)
Bumps ruff to latest. I'm not sure if we have specific concerns in which
case we can pin lower. Right now we are just behind a few.

### What
Upgrades ruff and runs `py-fmt` which handles almost all the fixes. A
few minor cleanups for the remaining.
Some of the biggest pieces are:
* [TC006](https://docs.astral.sh/ruff/rules/runtime-cast-value/)
*
[UP045](https://docs.astral.sh/ruff/rules/non-pep604-annotation-optional/)
* This was the motivating factor here since it kicked in for a new repo
but I knew we had inconsistencies here
*
[C420](https://docs.astral.sh/ruff/rules/unnecessary-dict-comprehension-for-iterable/)
* Improved
[PLW1514](https://docs.astral.sh/ruff/rules/unspecified-encoding/)

### TODO
* ~https://github.com/rerun-io/rerun/issues/10958~ apparently this PR
still works on 3.9