re_perf_telemetry crate: allow metric scraping (#11051)
- **Allow scraping of metrics (http endpoint).**

Previously metrics could only be pushed to an OTLP endpoint.
This change adds the ability to start a dedicated HTTP server
to expose a /metrics endpoint for Prometheus-style scraping.

This change does not enable a metrics endpoint in the rerun
viewer, but it allows doing so by calling `::start_metrics_server()`.

---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>