Re-apply some reverted things (#11067)
### Related
* https://github.com/rerun-io/rerun/pull/11065
* https://github.com/rerun-io/rerun/issues/11028

### What
We do not yet support migrating schemas on cloud
(https://github.com/rerun-io/rerun/issues/11028) so we had to revert
changing the datatype of `Blob` from `List<u8>` to `LargeBinary`.

However, in the revert we threw away a lot of good baby with the bath
water.

This PR selectively cherry-pick back a bunch of things that we do want
in, without actually changing any data formats.

This lets us:
* Get in some good stuff
* Reduce merge conflicts on other PRs to main
* Reduce future merge conflicts when we want to change the datatype
again

NOTE: everything in here has already been reviewed before


### TODO
* [x] `@rerun-bot full-check`