End-to-end out-of-band entry headers: kickoff (#11077)
This introduces gRPC headers as first class citizens of the Rerun Data
Protocol.

To demonstrate it in action, two endpoints have been ported to use them:
* `rerun.cloud.v1alpha1.RerunCloudService/ReadDatasetEntry`
* `rerun.cloud.v1alpha1.RerunCloudService/GetPartitionTableSchema`

Update: actually, a third surprise endpoint joined the party:
`/WriteChunks` was already header based.

There are a bunch of reasons why we'd want to leverage headers more.
They are explained in:
* Sibling: https://github.com/rerun-io/dataplatform/pull/1650