Split AnyValues into AnyValues and DynamicArchetype (#11045)
### Related
* #10879

### What
Right now AnyValues is ambiguous. If we have two AnyValues with unique
archetypes but the same component name we get collisions (since we don't
combine the archetype with the name like we do for built ins). This
deprecates the paths for archetypes to get attached to AnyValues (with
[follow up to remove](https://github.com/rerun-io/rerun/issues/10908) in
0.26). Then creates a `DynamicArchetype` where we are explicit that we
make the names unique like in built in types.

For both python and rust all the logic is in `DynamicArchetype` with
private hooks to create one without an Archetype that we expose as
`AnyValues`.