Refactor setting selection to always use system commands (#11135)
### Related

Part of #10866.

### What

To be able to handle selection updates in one place to update the viewer
url I made all instances of setting the selection to instead send a
`SystemCommand::SetSelection`. This also means we can make selection not
be in a mutex.

Part of this is also moving `ItemCollection` to `re_global_context` to
be able to send it in the `SetSelection` command.

### TODO
 - [x] implement the web url updating using these changes