Add Example Of LeRobot Dataloader (#11167)
Our docs currently aren't clear that we support LeRobot via a built in
dataloader. This adds a simple example of it and highlights the
integration. Hopefully it makes it more discoverable via search on our
doc page as well.

I also updated the github link for our existing demo that points to our
lerobot fork. It was previously pointing to main but I synced main to
latest upstream.