Fix sharing selection as a fragment (#11161)
### Related

 - Closes RR-2222
- Based on #11153 because I made some changes there that would conflict
with this otherwise.

### What

Fixes selection sharing, which before this was only applied too early to
actually affect the selection.

This adds the variant `SetUrlFragment` to `UiCommand` in
`crates/store/re_redap_client/src/grpc.rs`, which I'm pretty sure is a
breaking change (according to semver).

Also adds a system command `SetUrlFragment`, which creates other system
commands when invoked. Which might be the first system command which
does that? The system command loop looks like it's made to handle that
though.