Unify button handling in `LabelContent`, `PropertyContent` and `SectionHeader` (#11164)
### Related
* closes https://github.com/rerun-io/rerun/issues/6191
* closes https://github.com/rerun-io/rerun/issues/6203
* related to https://github.com/rerun-io/rerun/issues/10762 (RR-1532)

### What

Unify button handling via new `ItemButtons` struct and introduce ext
trait to unify convenience helper fns.

This changes the default behavior of the PropertyContent to only show
icon buttons on hover. In some places this may be desired (usually if
there are a lot of items with the same buttons (e.g. blueprint panel)),
sometimes not (e.g. if buttons are only on select items in a list, they
should always be visible for discoverability). We should check that this
should make sense for all property content buttons.