Set timeline with most events as default timeline (#11217)
### Related

* Closes
https://linear.app/rerun/issue/RR-2282/better-default-timeline-when-loading-mcap.
* Related: https://github.com/rerun-io/rerun/issues/10739

### What

This PR changes the `default_timeline` logic (when no timeline was
selected by the user) to always switch to the timeline with the most
events on it.

It also adds the number of events to the timeline selector in the time
panel.

> [!IMPORTANT]
> It would be very nice if we could right-align the subdued text. That
way it would be much easier to see, at a glance, which of the timelines
has the most events (because the text would be the longest). **If
someone knows how to do this quickly, please point out in the review.**

| dark | light |
|--------|--------|
| <img width="233" height="199" alt="image"
src="https://github.com/user-attachments/assets/2c3eacf0-2a5e-4ffb-b334-a8ba3af28159"
/> | <img width="233" height="201" alt="image"
src="https://github.com/user-attachments/assets/96d094cc-596d-4a2e-ae95-4a55201fe0b1"
/> |

---------

Co-authored-by: Andreas Reich <r_andreas2@web.de>